requirements.txt
python-dotenv
github3.py
Markdown
moment
ibm-watson
joblib
pandas
cloudant
